[WGSL] Using UINT_MAX for binding triggers a RELEASE_ASSERT https://bugs.webkit.org/show_bug.cgi?id=286695 rdar://143741471
Reviewed by Mike Wyrzykowski.

The `binding` value was used as the key to a HashMap, but the key had type `unsigned`
and `UINT_MAX` was used as the deleted value, so using `@binding(UINT_MAX)` triggered
a release assert in the HashMap.

* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
* Source/WebGPU/WGSL/tests/valid/binding-uint-max.wgsl: Added.

Canonical link: https://commits.webkit.org/289563@main

+++ Source/WebGPU/WGSL/GlobalVariableRewriter.cpp
@@ -78,7 +78,7 @@ private:
     };
 
     template<typename Value>
-    using IndexMap = HashMap<unsigned, Value, WTF::IntHash<unsigned>, WTF::UnsignedWithZeroKeyHashTraits<unsigned>>;
+    using IndexMap = HashMap<uint64_t, Value, WTF::IntHash<uint64_t>, WTF::UnsignedWithZeroKeyHashTraits<uint64_t>>;
 
     using UsedResources = IndexMap<IndexMap<Global*>>;
     using UsedPrivateGlobals = Vector<Global*>;
